\name{ask2loadFile}
\alias{ask2loadFile}
\title{ ask2loadFile }
\arguments{
\item{extension}{character string representing file extension}
}
\value{a list containing
\item{FileName}{ Dateiname der vom Benutzer ausgeaaehlt wurde}
\item{InDirectory}{ Directory welches vom Benutzer bestimmt wurde}

}
\description{Benutzerdialog zum Einstellen eines Directory und zum Vorbereiten des Lesens des entsprechenden DataIO Dateityps}

\author{Michael Thrun
}
