\name{fileparts}
\alias{fileparts}
\title{ fileparts }
\arguments{
\item{pathname}{character string representing pathname to be parsed}
}
\value{a list containing
\item{pathstr}{character string representing directory path}
\item{name}{character string representing base of file name}
\item{ext}{character string representing file extension}
}
\description{to be written}
\author{Michael Thrun}
