\name{OldData2NewData}
\alias{OldData2NewData}
\title{
  wandelt altes Format in neues Format automatisch um. 
}
\description{
Wandelt alte .data Daten in das neue .data Format um.
}
\usage{
OldData2NewData(Filename,Directory)
}
\arguments{
  \item{Filename}{String: Name der Datei die umgewandelt werden soll}
  \item{Directory}{String: Pfad des Ordners in dem die Datei liegt}
}
\value{ Rueckgabewert entspricht dem von WriteData fuer die neue .data Datei}
\author{Michael Thrun}
